Reducing Makefile
Now django settings is only set if it’s not empty! Which avoid errors
of not defined/exist settings file. We added the path and shell
helpers. We can easily launch ou django project with gunicorn and the
setup now accepts variable value to initial username/email to
superuser. The postgres OS X app is now only appended to PATH if OS X
is detected.